MovieRaterApi/requirements-dev.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.8.5 
python-decouple==3.5
pytz==2020.1
sqlparse==0.4.1  
static3==0.7.0
wincertstore==0.2
